Extracted GCEPD pv tests
